Added 2 new files from the SDL_gfx library: 	video/SDL_rotozoom.cpp video/SDL_rotozoom.h Also added a resizing method for KD_Font Corrected an awful memory leak. Now KD_Font has a destructor...

+++ krystaldrop/video/spriteinstance.cpp
@@ -64,7 +64,7 @@ bool KD_SpriteInstance::Display()
 			ret = true;
 		}
 
-		spr->Display(x,y, currentAnim, currentFrame);
+		spr->Display(x,y, currentAnim, (int)currentFrame);
 	}
 	else
 	{
@@ -86,7 +86,7 @@ bool KD_SpriteInstance::Display()
 			ret = true;
 		}
 
-		spr->Display(x,y, currentAnim, currentFrame);
+		spr->Display(x,y, currentAnim, (int)currentFrame);
 	}
 
 	return ret;
